refactor: small reorganisation of launchpad crate
The crate is slightly reorganised to accommodate the addition of the launcher binary. It will be
useful to have the launcher and TUI tightly coupled, with the same version number.

The changes:
* Move `node-launchpad` into its own `bin` directory. Shortly the launcher will be added as a
  directory alongside it.
* Since the `Cli` struct was the only thing in the `cli` module, the module can be removed in favour
  of the struct sitting alongside the TUI binary's `main` module.
* Correct misspelled version keys in references.
* Tidy dependencies to be sorted alphabetically for easier reference.
