+++ sn_node_launchpad/Cargo.toml
@@ -12,11 +12,9 @@ build = "build.rs"
 
 [[bin]]
 name = "node-launchpad"
-path = "src/main.rs"
+path = "src/bin/tui/main.rs"
 
 [dependencies]
-sn_peers_acquisition = { verison = "0.2.10", path = "../sn_peers_acquisition" }
-sn_service_management = { verison = "0.2.4", path = "../sn_service_management" }
 better-panic = "0.3.0"
 clap = { version = "4.4.5", features = [
     "derive",
@@ -40,9 +38,11 @@ log = "0.4.20"
 pretty_assertions = "1.4.0"
 ratatui = { version = "0.26.0", features = ["serde", "macros"] }
 serde = { version = "1.0.188", features = ["derive"] }
-sn-node-manager = { verison = "0.7.4", path = "../sn_node_manager" }
 serde_json = "1.0.107"
 signal-hook = "0.3.17"
+sn-node-manager = { version = "0.7.4", path = "../sn_node_manager" }
+sn_peers_acquisition = { version = "0.2.10", path = "../sn_peers_acquisition" }
+sn_service_management = { version = "0.2.4", path = "../sn_service_management" }
 strip-ansi-escapes = "0.2.0"
 strum = { version = "0.26.1", features = ["derive"] }
 tokio = { version = "1.32.0", features = ["full"] }
